requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
httpx==0.28.1
PyJWT==2.8.1
